Refresh carbuncle MP regen everytime a SMN has their turn checked
Implementing a "perfect" check for this requires us to check every character in turn order on every turn change, which is kind of a pain in the ass. Simulate by refreshing every time the SMN's turn is checked normally - on round changes, turn start, turn end.

This ought to be good enough; Swiftcast can only be used once per phase anyway.